test: move vec_progress unit tests out of the source file
# Summary

`vec_progress.rs` carried 919 lines of inline tests around 430 lines of
code; the tests now live in their own file, and `CLAUDE.md` records when
to do this.

# Details

The tests are white-box: they compare `VecProgress` against a reference
model by reading its internal entry vector. Making them a sibling module
under `progress/` would have required widening `quorum_set` and
`entries` to `pub(crate)`, so the file is instead a child module of
`vec_progress` itself, declared from within `vec_progress.rs`. That
layout already appears in the tree (`type_config.rs` with
`type_config/`) and keeps the private access the tests rely on without
loosening anything.

Measuring the tree first showed the two test layouts are not arbitrary:
inline modules run a median of 58 lines and 40 of the 59 are under 100,
while separate test files run a median of 137. The split tracks test
size, so `CLAUDE.md` documents that threshold rather than declaring one
layout correct, and directories holding a mix are called out as
expected. `vec_progress` was the clearest outlier at more than twice its
own source length.

Move only: the same 27 tests run and pass, and no test body changed.
